pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoforge"
version = "1.0.0"
description = "Autonomous AI Engineering Orchestrator for GitLab"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["ai", "devops", "gitlab", "agents", "orchestration", "cicd"]

authors = [
    {name = "AutoForge Team"}
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "aiohttp>=3.11.0",
    "anthropic>=0.42.0",
    "python-gitlab>=5.0.0",
    "python-json-logger>=3.2.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
full = [
    "celery[redis]>=5.4.0",
    "redis>=5.2.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "chromadb>=0.5.23",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
tracing = [
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.urls]
Homepage = "https://github.com/DiveshK007/AutoForge"
Repository = "https://github.com/DiveshK007/AutoForge"
Documentation = "https://github.com/DiveshK007/AutoForge/tree/main/docs"
Issues = "https://github.com/DiveshK007/AutoForge/issues"

[project.scripts]
autoforge = "backend.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

# ─── Tool Configuration ─────────────────────────────────

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit (security)
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # Ruff-specific rules
]
ignore = [
    "S101",  # assert used (fine for tests)
    "T201",  # print found (we use structlog mostly)
    "B008",  # function call in default arg (FastAPI Depends)
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"backend/tests/**" = ["S101", "S106"]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
